refactor: SlackService 템플릿콜백 패턴 사용하여 리팩터링
