+++ backend/src/main/java/com/woowacourse/zzimkkong/controller/ManagerReservationController.java
@@ -2,12 +2,11 @@ package com.woowacourse.zzimkkong.controller;
 
 import com.woowacourse.zzimkkong.domain.LoginEmail;
 import com.woowacourse.zzimkkong.config.logaspect.LogMethodExecutionTime;
-import com.woowacourse.zzimkkong.domain.Member;
 import com.woowacourse.zzimkkong.dto.reservation.*;
 import com.woowacourse.zzimkkong.dto.slack.SlackResponse;
 import com.woowacourse.zzimkkong.dto.member.LoginEmailDto;
 import com.woowacourse.zzimkkong.service.ReservationService;
-import com.woowacourse.zzimkkong.service.SlackService;
+import com.woowacourse.zzimkkong.service.slack.SlackService;
 import com.woowacourse.zzimkkong.service.strategy.ManagerReservationStrategy;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
